requirements.txt
# =============================================================================
# AGRO-CLIMATIC DATA WAREHOUSE - REQUIREMENTS
# =============================================================================
# Python version: 3.11+
# Last updated: 2024
# =============================================================================

# -----------------------------------------------------------------------------
# CORE DATA PROCESSING
# -----------------------------------------------------------------------------
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2                    # Parquet support for data lake
python-dateutil==2.8.2

# -----------------------------------------------------------------------------
# DATABASE & STORAGE
# -----------------------------------------------------------------------------
psycopg2-binary==2.9.9             # PostgreSQL adapter
SQLAlchemy==2.0.25                 # ORM and SQL toolkit
alembic==1.13.1                    # Database migrations
redis==5.0.1                       # Caching layer (optional)

# -----------------------------------------------------------------------------
# API & HTTP CLIENTS
# -----------------------------------------------------------------------------
requests==2.31.0
httpx==0.26.0                      # Async HTTP client
urllib3==2.1.0
tenacity==8.2.3                    # Retry logic
ratelimit==2.2.1                   # Rate limiting decorators

# -----------------------------------------------------------------------------
# WEB SCRAPING
# -----------------------------------------------------------------------------
beautifulsoup4==4.12.2
lxml==5.1.0                        # Fast XML/HTML parser
html5lib==1.1                      # lenient HTML parser
scrapy==2.11.0                     # Advanced scraping framework (optional)
selenium==4.16.2                   # Browser automation (for JS sites)
webdriver-manager==4.0.1           # Chrome driver management

# -----------------------------------------------------------------------------
# NATURAL LANGUAGE PROCESSING
# -----------------------------------------------------------------------------
spacy==3.7.2
# Run after install: python -m spacy download en_core_web_sm
# Run after install: python -m spacy download en_core_web_trf (for transformers)

# Transformer models (optional, for advanced NLP)
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Text processing
regex==2023.12.25
nltk==3.8.1                        # Natural language toolkit
textblob==0.17.1                   # Simple NLP API

# -----------------------------------------------------------------------------
# DATA VALIDATION & QUALITY
# -----------------------------------------------------------------------------
pydantic==2.5.3                    # Data validation using Python type hints
pydantic-settings==2.1.0           # Settings management
jsonschema==4.20.0                 # JSON Schema validation
great-expectations==0.18.8         # Data quality framework (optional)
cerberus==1.3.5                    # Lightweight validation

# -----------------------------------------------------------------------------
# CONFIGURATION & ENVIRONMENT
# -----------------------------------------------------------------------------
python-dotenv==1.0.0               # Environment variable management
pyyaml==6.0.1                      # YAML parsing
toml==0.10.2                       # TOML parsing
click==8.1.7                       # CLI framework
typer==0.9.0                       # Modern CLI (alternative to click)

# -----------------------------------------------------------------------------
# LOGGING & MONITORING
# -----------------------------------------------------------------------------
structlog==23.2.0                  # Structured logging
loguru==0.7.2                      # Modern logging (alternative)
prometheus-client==0.19.0          # Metrics for monitoring

# -----------------------------------------------------------------------------
# TESTING
# -----------------------------------------------------------------------------
pytest==7.4.4                      # Testing framework
pytest-cov==4.1.0                  # Coverage reporting
pytest-asyncio==0.23.3             # Async test support
pytest-mock==3.12.0                # Mocking support
factory-boy==3.3.0                 # Test data factories
faker==22.0.0                      # Generate fake data
responses==0.24.1                  # Mock HTTP requests
httmock==1.4.0                     # HTTP mocking (alternative)

# -----------------------------------------------------------------------------
# DEVELOPMENT & CODE QUALITY
# -----------------------------------------------------------------------------
black==23.12.1                     # Code formatting
isort==5.13.2                      # Import sorting
ruff==0.1.9                        # Fast Python linter
mypy==1.8.0                        # Static type checking
pylint==3.0.3                      # Code analysis
bandit==1.7.6                      # Security linting
pre-commit==3.6.0                  # Git hooks

# -----------------------------------------------------------------------------
# DOCUMENTATION
# -----------------------------------------------------------------------------
sphinx==7.2.6                      # Documentation generator
sphinx-rtd-theme==2.0.0            # ReadTheDocs theme
mkdocs==1.5.3                      # Project documentation
mkdocs-material==9.5.3             # Material theme for MkDocs

# -----------------------------------------------------------------------------
# NOTEBOOKS & VISUALIZATION
# -----------------------------------------------------------------------------
jupyter==1.0.0
ipython==8.19.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0                     # Interactive charts
folium==0.15.1                     # Geographic visualization
geopandas==0.14.2                  # Geospatial data handling

# -----------------------------------------------------------------------------
# UTILITIES
# -----------------------------------------------------------------------------
tqdm==4.66.1                       # Progress bars
joblib==1.3.2                      # Parallel processing
cachetools==5.3.2                  # Caching utilities
diskcache==5.6.3                   # Disk-based caching
python-slugify==8.0.1              # URL-friendly slugs

# -----------------------------------------------------------------------------
# ORCHESTRATION (Future: Airflow)
# -----------------------------------------------------------------------------
# apache-airflow==2.8.0            # Uncomment when adding Airflow
# celery==5.3.6                    # Distributed task queue

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  (Linux/Mac)
#    venv\Scripts\activate     (Windows)
#
# 2. Install requirements:
#    pip install -r requirements.txt
#
# 3. Download spaCy models:
#    python -m spacy download en_core_web_sm
#    python -m spacy download en_core_web_md  (larger model, better accuracy)
#
# 4. For GPU support (PyTorch):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 5. Verify installation:
#    python -c "import pandas, spacy, psycopg2; print('All imports successful')"
#
# =============================================================================